test: cleanup conftest
